+++ skin-support/src/main/java/skin/support/widget/SkinCompatCheckedTextView.java
@@ -41,7 +41,7 @@ public class SkinCompatCheckedTextView extends AppCompatCheckedTextView implemen
         }
         mBackgroundTintHelper = new SkinCompatBackgroundHelper(this);
         mBackgroundTintHelper.loadFromAttributes(attrs, defStyleAttr);
-        mTextHelper = new SkinCompatTextHelper(this);
+        mTextHelper = SkinCompatTextHelper.create(this);
         mTextHelper.loadFromAttributes(attrs, defStyleAttr);
 
         TintTypedArray a = TintTypedArray.obtainStyledAttributes(getContext(), attrs,
@@ -63,6 +63,11 @@ public class SkinCompatCheckedTextView extends AppCompatCheckedTextView implemen
         if (mBackgroundTintHelper != null) {
             mBackgroundTintHelper.onSetBackgroundResource(resId);
         }
+    }
+
+    @Override
+    public void setTextAppearance(int resId) {
+        setTextAppearance(getContext(), resId);
     }
 
     @Override