+++ skin-support/src/main/java/skin/support/widget/SkinCompatEditText.java
@@ -33,7 +33,7 @@ public class SkinCompatEditText extends AppCompatEditText implements SkinCompatS
         }
         mBackgroundTintHelper = new SkinCompatBackgroundHelper(this);
         mBackgroundTintHelper.loadFromAttributes(attrs, defStyleAttr);
-        mTextHelper = new SkinCompatTextHelper(this);
+        mTextHelper = SkinCompatTextHelper.create(this);
         mTextHelper.loadFromAttributes(attrs, defStyleAttr);
     }
 
@@ -43,6 +43,11 @@ public class SkinCompatEditText extends AppCompatEditText implements SkinCompatS
         if (mBackgroundTintHelper != null) {
             mBackgroundTintHelper.onSetBackgroundResource(resId);
         }
+    }
+
+    @Override
+    public void setTextAppearance(int resId) {
+        setTextAppearance(getContext(), resId);
     }
 
     @Override