+++ skin-support/src/main/java/skin/support/widget/SkinCompatTextHelper.java
@@ -2,6 +2,8 @@ package skin.support.widget;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
+import android.graphics.drawable.Drawable;
+import android.os.Build;
 import android.support.v7.widget.TintTypedArray;
 import android.util.AttributeSet;
 import android.widget.TextView;
@@ -17,10 +19,21 @@ import skin.support.utils.SkinLog;
 public class SkinCompatTextHelper extends SkinCompatHelper {
     private static final String TAG = SkinCompatTextHelper.class.getSimpleName();
 
-    private final TextView mView;
+    public static SkinCompatTextHelper create(TextView textView) {
+        if (Build.VERSION.SDK_INT >= 17) {
+            return new SkinCompatTextHelperV17(textView);
+        }
+        return new SkinCompatTextHelper(textView);
+    }
+
+    final TextView mView;
 
     private int mTextColorResId = INVALID_ID;
     private int mTextColorHintResId = INVALID_ID;
+    protected int mDrawableBottomResId = INVALID_ID;
+    protected int mDrawableLeftResId = INVALID_ID;
+    protected int mDrawableRightResId = INVALID_ID;
+    protected int mDrawableTopResId = INVALID_ID;
 
     public SkinCompatTextHelper(TextView view) {
         mView = view;
@@ -34,6 +47,19 @@ public class SkinCompatTextHelper extends SkinCompatHelper {
                 R.styleable.SkinCompatTextHelper, defStyleAttr, 0);
         final int ap = a.getResourceId(R.styleable.SkinCompatTextHelper_android_textAppearance, INVALID_ID);
         SkinLog.d(TAG, "ap = " + ap);
+
+        if (a.hasValue(R.styleable.AppCompatTextHelper_android_drawableLeft)) {
+            mDrawableLeftResId = a.getResourceId(R.styleable.AppCompatTextHelper_android_drawableLeft, INVALID_ID);
+        }
+        if (a.hasValue(R.styleable.AppCompatTextHelper_android_drawableTop)) {
+            mDrawableTopResId = a.getResourceId(R.styleable.AppCompatTextHelper_android_drawableTop, INVALID_ID);
+        }
+        if (a.hasValue(R.styleable.AppCompatTextHelper_android_drawableRight)) {
+            mDrawableRightResId = a.getResourceId(R.styleable.AppCompatTextHelper_android_drawableRight, INVALID_ID);
+        }
+        if (a.hasValue(R.styleable.AppCompatTextHelper_android_drawableBottom)) {
+            mDrawableBottomResId = a.getResourceId(R.styleable.AppCompatTextHelper_android_drawableBottom, INVALID_ID);
+        }
         a.recycle();
 
         if (ap != INVALID_ID) {
@@ -78,7 +104,45 @@ public class SkinCompatTextHelper extends SkinCompatHelper {
             SkinLog.d(TAG, "mTextColorHintResId = " + mTextColorHintResId);
         }
         a.recycle();
-        applySkin();
+        applyTextColorResource();
+        applyTextColorHintResource();
+    }
+
+    private void applyTextColorHintResource() {
+        mTextColorHintResId = checkResourceId(mTextColorHintResId);
+        if (mTextColorHintResId != INVALID_ID) {
+            ColorStateList color = SkinCompatResources.getInstance().getColorStateList(mTextColorHintResId);
+            mView.setHintTextColor(color);
+        }
+    }
+
+    private void applyTextColorResource() {
+        mTextColorResId = checkResourceId(mTextColorResId);
+        if (mTextColorResId != INVALID_ID) {
+            ColorStateList color = SkinCompatResources.getInstance().getColorStateList(mTextColorResId);
+            mView.setTextColor(color);
+        }
+    }
+
+    protected void applyCompoundDrawablesResource() {
+        Drawable drawableLeft = null, drawableTop = null, drawableRight = null, drawableBottom = null;
+        mDrawableLeftResId = checkResourceId(mDrawableLeftResId);
+        if (mDrawableLeftResId != INVALID_ID) {
+            drawableLeft = SkinCompatResources.getInstance().getDrawable(mDrawableLeftResId);
+        }
+        mDrawableTopResId = checkResourceId(mDrawableTopResId);
+        if (mDrawableTopResId != INVALID_ID) {
+            drawableTop = SkinCompatResources.getInstance().getDrawable(mDrawableTopResId);
+        }
+        mDrawableRightResId = checkResourceId(mDrawableRightResId);
+        if (mDrawableRightResId != INVALID_ID) {
+            drawableRight = SkinCompatResources.getInstance().getDrawable(mDrawableRightResId);
+        }
+        mDrawableBottomResId = checkResourceId(mDrawableBottomResId);
+        if (mDrawableBottomResId != INVALID_ID) {
+            drawableBottom = SkinCompatResources.getInstance().getDrawable(mDrawableBottomResId);
+        }
+        mView.setCompoundDrawablesWithIntrinsicBounds(drawableLeft, drawableTop, drawableRight, drawableBottom);
     }
 
     public int getTextColorResId() {
@@ -86,15 +150,8 @@ public class SkinCompatTextHelper extends SkinCompatHelper {
     }
 
     public void applySkin() {
-        mTextColorResId = checkResourceId(mTextColorResId);
-        if (mTextColorResId != INVALID_ID) {
-            ColorStateList color = SkinCompatResources.getInstance().getColorStateList(mTextColorResId);
-            mView.setTextColor(color);
-        }
-        mTextColorHintResId = checkResourceId(mTextColorHintResId);
-        if (mTextColorHintResId != INVALID_ID) {
-            ColorStateList color = SkinCompatResources.getInstance().getColorStateList(mTextColorHintResId);
-            mView.setHintTextColor(color);
-        }
+        applyTextColorResource();
+        applyTextColorHintResource();
+        applyCompoundDrawablesResource();
     }
 }