+++ skin-support/src/main/java/skin/support/widget/SkinCompatTextView.java
@@ -30,7 +30,7 @@ public class SkinCompatTextView extends AppCompatTextView implements SkinCompatS
         }
         mBackgroundTintHelper = new SkinCompatBackgroundHelper(this);
         mBackgroundTintHelper.loadFromAttributes(attrs, defStyleAttr);
-        mTextHelper = new SkinCompatTextHelper(this);
+        mTextHelper = SkinCompatTextHelper.create(this);
         mTextHelper.loadFromAttributes(attrs, defStyleAttr);
     }
 
@@ -40,6 +40,11 @@ public class SkinCompatTextView extends AppCompatTextView implements SkinCompatS
         if (mBackgroundTintHelper != null) {
             mBackgroundTintHelper.onSetBackgroundResource(resId);
         }
+    }
+
+    @Override
+    public void setTextAppearance(int resId) {
+        setTextAppearance(getContext(), resId);
     }
 
     @Override