+++ ChangeLog.md
@@ -20,35 +20,127 @@
 
 * [![v1.0.1](https://img.shields.io/badge/skin--support-v1.0.1-green.svg)](http://jcenter.bintray.com/skin/support/skin-support/1.0.1/) 支持所有基础控件换肤 依赖 appcompat-v7:25.1.0
   * View
+    * android:background
   * Button
+    * android:background
+    * android:textColor
+    * android:textColorHint
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * Spinner
+    * android:popupBackground
+    * android:background
   * SeekBar
+    * android:thumb
+    * android:indeterminateDrawable
+    * android:progressDrawable
   * CheckBox
+    * android:button
   * EditText
+    * android:background
+    * android:textColor
+    * android:textColorHint
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * TextView
+    * android:background
+    * android:textColor
+    * android:textColorHint
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * RatingBar
+    * android:indeterminateDrawable
+    * android:progressDrawable
   * ImageView
+    * android:background
+    * android:src
   * ProgressBar
+    * android:indeterminateDrawable
+    * android:progressDrawable
   * ImageButton
+    * android:background
+    * android:src
   * RadioButton
+    * android:button
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * FrameLayout
+    * android:background
   * LinearLayout
+    * android:background
   * RelativeLayout
+    * android:background
   * CheckedTextView
+    * android:checkMark
+    * android:background
+    * android:textColor
+    * android:textColorHint
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * AutoCompleteTextView
+    * android:popupBackground
+    * android:background
+    * android:textColor
+    * android:textColorHint
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * MultiAutoCompleteTextView
+    * android:popupBackground
+    * android:background
+    * android:textColor
+    * android:textColorHint
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * android.support.v7.widget.Toolbar
+    * android:background
+    * app:navigationIcon
+    * app:titleTextColor
+    * app:subtitleTextColor
 
 ## skin-support-design: material design 支持
 
 
 * [![v1.0.0](https://img.shields.io/badge/skin--support--design-v1.0.0-green.svg)](http://jcenter.bintray.com/skin/support/skin-support-design/1.0.0/) 新功能开发 依赖skin-support:1.1.2 design:25.1.0
   * BottomNavigationView
+    * app:itemIconTint
+    * app:itemTextColor
   * CollapsingToolbarLayout
+    * android:background
+    * app:contentScrim
+    * app:statusBarScrim
   * CoordinatorLayout
+    * android:background
   * FloatingActionButton
+    * android:src
+    * app:backgroundTint
+    * app:rippleColor
   * TextInputEditText
+    * android:background
+    * android:textColor
+    * android:textColorHint
+    * android:drawableLeft
+    * android:drawableTop
+    * android:drawableRight
+    * android:drawableBottom
   * TextInputLayout
+    * android:background
+    * android:textColorHint
+    * app:errorTextAppearance
+    * app:counterTextAppearance
 
 * [![v0.1.0](https://img.shields.io/badge/skin--support--design-v0.1.0-green.svg)](http://jcenter.bintray.com/skin/support/skin-support-design/0.1.0/) 新功能开发 依赖skin-support:1.1.0 design:25.1.0
 
@@ -56,11 +148,17 @@
 
 * [![v0.0.2](https://img.shields.io/badge/skin--support--design-v0.0.2-green.svg)](http://jcenter.bintray.com/skin/support/skin-support-design/0.0.2/) 支持以下三个控件换肤 依赖skin-support:1.0.2 design:25.1.0
   * TabLayout
+    * app:tabIndicatorColor
+    * app:tabSelectedTextColor
+    * app:tabTextColor
   * AppBarLayout
+    * android:background
   * NavigationView
+    * android:background
 
 ## skin-support-cardview: CardView 支持
 
 
 * [![v1.0.0](https://img.shields.io/badge/skin--support--cardview-v1.0.0-green.svg)](http://jcenter.bintray.com/skin/support/skin-support-cardview/1.0.0/) CardView 支持 依赖skin-support:1.1.2 cardview-v7:25.1.0
   * CardView
+    * app:cardBackgroundColor